🚅 perfs: split help function to psbt dir
