+++ src/psbt.js
@@ -2,18 +2,23 @@
 Object.defineProperty(exports, '__esModule', { value: true });
 exports.Psbt = void 0;
 const bip174_1 = require('bip174');
-const varuint = require('bip174/src/lib/converter/varint');
 const utils_1 = require('bip174/src/lib/utils');
 const address_1 = require('./address');
 const bufferutils_1 = require('./bufferutils');
 const networks_1 = require('./networks');
 const payments = require('./payments');
-const bip341_1 = require('./payments/bip341');
 const bscript = require('./script');
 const transaction_1 = require('./transaction');
 const bip371_1 = require('./psbt/bip371');
 const psbtutils_1 = require('./psbt/psbtutils');
 const transaction_2 = require('./psbt/transaction');
+const cache_1 = require('./psbt/global/cache');
+const global_1 = require('./psbt/global');
+const input_1 = require('./psbt/input');
+const sign_1 = require('./psbt/global/sign');
+const script_1 = require('./psbt/input/script');
+const output_1 = require('./psbt/output');
+const hash_1 = require('./psbt/global/hash');
 /**
  * These are the default arguments for a Psbt instance.
  */
@@ -82,7 +87,7 @@ class Psbt {
       transaction_2.transactionFromBuffer,
     );
     const psbt = new Psbt(opts, psbtBase);
-    checkTxForDupeIns(psbt.__CACHE.__TX, psbt.__CACHE);
+    (0, input_1.checkTxForDupeIns)(psbt.__CACHE.__TX, psbt.__CACHE);
     return psbt;
   }
   constructor(
@@ -171,7 +176,7 @@ class Psbt {
   }
   setVersion(version) {
     check32Bit(version);
-    checkInputsForPartialSig(this.data.inputs, 'setVersion');
+    (0, input_1.checkInputsForPartialSig)(this.data.inputs, 'setVersion');
     const c = this.__CACHE;
     c.__TX.version = version;
     c.__EXTRACTED_TX = undefined;
@@ -179,7 +184,7 @@ class Psbt {
   }
   setLocktime(locktime) {
     check32Bit(locktime);
-    checkInputsForPartialSig(this.data.inputs, 'setLocktime');
+    (0, input_1.checkInputsForPartialSig)(this.data.inputs, 'setLocktime');
     const c = this.__CACHE;
     c.__TX.locktime = locktime;
     c.__EXTRACTED_TX = undefined;
@@ -187,7 +192,7 @@ class Psbt {
   }
   setInputSequence(inputIndex, sequence) {
     check32Bit(sequence);
-    checkInputsForPartialSig(this.data.inputs, 'setInputSequence');
+    (0, input_1.checkInputsForPartialSig)(this.data.inputs, 'setInputSequence');
     const c = this.__CACHE;
     if (c.__TX.ins.length <= inputIndex) {
       throw new Error('Input index too high');
@@ -213,16 +218,17 @@ class Psbt {
       );
     }
     (0, bip371_1.checkTaprootInputFields)(inputData, inputData, 'addInput');
-    checkInputsForPartialSig(this.data.inputs, 'addInput');
-    if (inputData.witnessScript) checkInvalidP2WSH(inputData.witnessScript);
+    (0, input_1.checkInputsForPartialSig)(this.data.inputs, 'addInput');
+    if (inputData.witnessScript)
+      (0, script_1.checkInvalidP2WSH)(inputData.witnessScript);
     const c = this.__CACHE;
     this.data.addInput(inputData);
     const txIn = c.__TX.ins[c.__TX.ins.length - 1];
-    checkTxInputCache(c, txIn);
+    (0, cache_1.checkTxInputCache)(c, txIn);
     const inputIndex = this.data.inputs.length - 1;
     const input = this.data.inputs[inputIndex];
     if (input.nonWitnessUtxo) {
-      addNonWitnessTxCache(this.__CACHE, input, inputIndex);
+      (0, cache_1.addNonWitnessTxCache)(this.__CACHE, input, inputIndex);
     }
     c.__FEE = undefined;
     c.__FEE_RATE = undefined;
@@ -245,7 +251,7 @@ class Psbt {
           `Requires single object with at least [script or address] and [value]`,
       );
     }
-    checkInputsForPartialSig(this.data.inputs, 'addOutput');
+    (0, input_1.checkInputsForPartialSig)(this.data.inputs, 'addOutput');
     const { address } = outputData;
     if (typeof address === 'string') {
       const { network } = this.opts;
@@ -261,18 +267,19 @@ class Psbt {
     return this;
   }
   extractTransaction(disableFeeCheck) {
-    if (!this.data.inputs.every(isFinalized)) throw new Error('Not finalized');
+    if (!this.data.inputs.every(global_1.isFinalized))
+      throw new Error('Not finalized');
     const c = this.__CACHE;
     if (!disableFeeCheck) {
       checkFees(this, c, this.opts);
     }
     if (c.__EXTRACTED_TX) return c.__EXTRACTED_TX;
     const tx = c.__TX.clone();
-    inputFinalizeGetAmts(this.data.inputs, tx, c, true);
+    (0, input_1.inputFinalizeGetAmts)(this.data.inputs, tx, c, true);
     return tx;
   }
   getFeeRate() {
-    return getTxCacheValue(
+    return (0, cache_1.getTxCacheValue)(
       '__FEE_RATE',
       'fee rate',
       this.data.inputs,
@@ -280,7 +287,12 @@ class Psbt {
     );
   }
   getFee() {
-    return getTxCacheValue('__FEE', 'fee', this.data.inputs, this.__CACHE);
+    return (0, cache_1.getTxCacheValue)(
+      '__FEE',
+      'fee',
+      this.data.inputs,
+      this.__CACHE,
+    );
   }
   finalizeAllInputs() {
     (0, utils_1.checkForInput)(this.data.inputs, 0); // making sure we have at least one
@@ -313,14 +325,18 @@ class Psbt {
       );
     throw new Error(`Cannot finalize input #${inputIndex}. Not Taproot.`);
   }
-  _finalizeInput(inputIndex, input, finalScriptsFunc = getFinalScripts) {
+  _finalizeInput(
+    inputIndex,
+    input,
+    finalScriptsFunc = global_1.getFinalScripts,
+  ) {
     const { script, isP2SH, isP2WSH, isSegwit } = getScriptFromInput(
       inputIndex,
       input,
       this.__CACHE,
     );
     if (!script) throw new Error(`No script found for input #${inputIndex}`);
-    checkPartialSigSighashes(input);
+    (0, sign_1.checkPartialSigSighashes)(input);
     const { finalScriptSig, finalScriptWitness } = finalScriptsFunc(
       inputIndex,
       input,
@@ -370,8 +386,12 @@ class Psbt {
   }
   getInputType(inputIndex) {
     const input = (0, utils_1.checkForInput)(this.data.inputs, inputIndex);
-    const script = getScriptFromUtxo(inputIndex, input, this.__CACHE);
-    const result = getMeaningfulScript(
+    const script = (0, script_1.getScriptFromUtxo)(
+      inputIndex,
+      input,
+      this.__CACHE,
+    );
+    const result = (0, script_1.getMeaningfulScript)(
       script,
       inputIndex,
       'input',
@@ -380,7 +400,7 @@ class Psbt {
         redeemFromFinalWitnessScript(input.finalScriptWitness),
     );
     const type = result.type === 'raw' ? '' : result.type + '-';
-    const mainType = classifyScript(result.meaningfulScript);
+    const mainType = (0, script_1.classifyScript)(result.meaningfulScript);
     return type + mainType;
   }
   inputHasPubkey(inputIndex, pubkey) {
@@ -389,18 +409,23 @@ class Psbt {
   }
   inputHasHDKey(inputIndex, root) {
     const input = (0, utils_1.checkForInput)(this.data.inputs, inputIndex);
-    const derivationIsMine = bip32DerivationIsMine(root);
+    const derivationIsMine = (0, global_1.bip32DerivationIsMine)(root);
     return (
       !!input.bip32Derivation && input.bip32Derivation.some(derivationIsMine)
     );
   }
   outputHasPubkey(outputIndex, pubkey) {
     const output = (0, utils_1.checkForOutput)(this.data.outputs, outputIndex);
-    return pubkeyInOutput(pubkey, output, outputIndex, this.__CACHE);
+    return (0, output_1.pubkeyInOutput)(
+      pubkey,
+      output,
+      outputIndex,
+      this.__CACHE,
+    );
   }
   outputHasHDKey(outputIndex, root) {
     const output = (0, utils_1.checkForOutput)(this.data.outputs, outputIndex);
-    const derivationIsMine = bip32DerivationIsMine(root);
+    const derivationIsMine = (0, global_1.bip32DerivationIsMine)(root);
     return (
       !!output.bip32Derivation && output.bip32Derivation.some(derivationIsMine)
     );
@@ -441,7 +466,7 @@ class Psbt {
       const sig = bscript.signature.decode(pSig.signature);
       const { hash, script } =
         sighashCache !== sig.hashType
-          ? getHashForSig(
+          ? (0, hash_1.getHashForSig)(
               inputIndex,
               Object.assign({}, input, { sighashType: sig.hashType }),
               this.__CACHE,
@@ -451,7 +476,7 @@ class Psbt {
       sighashCache = sig.hashType;
       hashCache = hash;
       scriptCache = script;
-      checkScriptForPubkey(pSig.pubkey, script, 'verify');
+      (0, script_1.checkScriptForPubkey)(pSig.pubkey, script, 'verify');
       results.push(validator(pSig.pubkey, hash, sig.signature));
     }
     return results.every(res => res === true);
@@ -466,14 +491,14 @@ class Psbt {
       throw new Error('Need validator function to validate signatures');
     pubkey = pubkey && (0, bip371_1.toXOnly)(pubkey);
     const allHashses = pubkey
-      ? getTaprootHashesForSig(
+      ? (0, hash_1.getTaprootHashesForSig)(
           inputIndex,
           input,
           this.data.inputs,
           pubkey,
           this.__CACHE,
         )
-      : getAllTaprootHashesForSig(
+      : (0, hash_1.getAllTaprootHashesForSig)(
           inputIndex,
           input,
           this.data.inputs,
@@ -673,7 +698,7 @@ class Psbt {
     keyPair,
     sighashTypes = [transaction_1.Transaction.SIGHASH_ALL],
   ) {
-    const { hash, sighashType } = getHashAndSighashType(
+    const { hash, sighashType } = (0, hash_1.getHashAndSighashType)(
       this.data.inputs,
       inputIndex,
       keyPair.publicKey,
@@ -766,7 +791,7 @@ class Psbt {
     keyPair,
     sighashTypes = [transaction_1.Transaction.SIGHASH_ALL],
   ) {
-    const { hash, sighashType } = getHashAndSighashType(
+    const { hash, sighashType } = (0, hash_1.getHashAndSighashType)(
       this.data.inputs,
       inputIndex,
       keyPair.publicKey,
@@ -848,7 +873,7 @@ class Psbt {
       throw new Error(
         `Need Schnorr Signer to sign taproot input #${inputIndex}.`,
       );
-    const hashesForSig = getTaprootHashesForSig(
+    const hashesForSig = (0, hash_1.getTaprootHashesForSig)(
       inputIndex,
       input,
       this.data.inputs,
@@ -866,15 +891,15 @@ class Psbt {
     return hashesForSig;
   }
   toBuffer() {
-    checkCache(this.__CACHE);
+    (0, cache_1.checkCache)(this.__CACHE);
     return this.data.toBuffer();
   }
   toHex() {
-    checkCache(this.__CACHE);
+    (0, cache_1.checkCache)(this.__CACHE);
     return this.data.toHex();
   }
   toBase64() {
-    checkCache(this.__CACHE);
+    (0, cache_1.checkCache)(this.__CACHE);
     return this.data.toBase64();
   }
   updateGlobal(updateData) {
@@ -882,7 +907,8 @@ class Psbt {
     return this;
   }
   updateInput(inputIndex, updateData) {
-    if (updateData.witnessScript) checkInvalidP2WSH(updateData.witnessScript);
+    if (updateData.witnessScript)
+      (0, script_1.checkInvalidP2WSH)(updateData.witnessScript);
     (0, bip371_1.checkTaprootInputFields)(
       this.data.inputs[inputIndex],
       updateData,
@@ -890,7 +916,7 @@ class Psbt {
     );
     this.data.updateInput(inputIndex, updateData);
     if (updateData.nonWitnessUtxo) {
-      addNonWitnessTxCache(
+      (0, cache_1.addNonWitnessTxCache)(
         this.__CACHE,
         this.data.inputs[inputIndex],
         inputIndex,
@@ -926,50 +952,6 @@ class Psbt {
   }
 }
 exports.Psbt = Psbt;
-function canFinalize(input, script, scriptType) {
-  switch (scriptType) {
-    case 'pubkey':
-    case 'pubkeyhash':
-    case 'witnesspubkeyhash':
-      return hasSigs(1, input.partialSig);
-    case 'multisig':
-      const p2ms = payments.p2ms({ output: script });
-      return hasSigs(p2ms.m, input.partialSig, p2ms.pubkeys);
-    default:
-      return false;
-  }
-}
-function checkCache(cache) {
-  if (cache.__UNSAFE_SIGN_NONSEGWIT !== false) {
-    throw new Error('Not BIP174 compliant, can not export');
-  }
-}
-function hasSigs(neededSigs, partialSig, pubkeys) {
-  if (!partialSig) return false;
-  let sigs;
-  if (pubkeys) {
-    sigs = pubkeys
-      .map(pkey => {
-        const pubkey = compressPubkey(pkey);
-        return partialSig.find(pSig => pSig.pubkey.equals(pubkey));
-      })
-      .filter(v => !!v);
-  } else {
-    sigs = partialSig;
-  }
-  if (sigs.length > neededSigs) throw new Error('Too many signatures');
-  return sigs.length === neededSigs;
-}
-function isFinalized(input) {
-  return !!input.finalScriptSig || !!input.finalScriptWitness;
-}
-function bip32DerivationIsMine(root) {
-  return d => {
-    if (!d.masterFingerprint.equals(root.fingerprint)) return false;
-    if (!root.derivePath(d.path).publicKey.equals(d.pubkey)) return false;
-    return true;
-  };
-}
 function check32Bit(num) {
   if (
     typeof num !== 'number' ||
@@ -994,362 +976,8 @@ function checkFees(psbt, cache, opts) {
     );
   }
 }
-function checkInputsForPartialSig(inputs, action) {
-  inputs.forEach(input => {
-    const throws = (0, bip371_1.isTaprootInput)(input)
-      ? (0, bip371_1.checkTaprootInputForSigs)(input, action)
-      : (0, psbtutils_1.checkInputForSig)(input, action);
-    if (throws)
-      throw new Error('Can not modify transaction, signatures exist.');
-  });
-}
-function checkPartialSigSighashes(input) {
-  if (!input.sighashType || !input.partialSig) return;
-  const { partialSig, sighashType } = input;
-  partialSig.forEach(pSig => {
-    const { hashType } = bscript.signature.decode(pSig.signature);
-    if (sighashType !== hashType) {
-      throw new Error('Signature sighash does not match input sighash type');
-    }
-  });
-}
-function checkScriptForPubkey(pubkey, script, action) {
-  if (!(0, psbtutils_1.pubkeyInScript)(pubkey, script)) {
-    throw new Error(
-      `Can not ${action} for this input with the key ${pubkey.toString('hex')}`,
-    );
-  }
-}
-function checkTxForDupeIns(tx, cache) {
-  tx.ins.forEach(input => {
-    checkTxInputCache(cache, input);
-  });
-}
-function checkTxInputCache(cache, input) {
-  const key =
-    (0, bufferutils_1.reverseBuffer)(Buffer.from(input.hash)).toString('hex') +
-    ':' +
-    input.index;
-  if (cache.__TX_IN_CACHE[key]) throw new Error('Duplicate input detected.');
-  cache.__TX_IN_CACHE[key] = 1;
-}
-function scriptCheckerFactory(payment, paymentScriptName) {
-  return (inputIndex, scriptPubKey, redeemScript, ioType) => {
-    const redeemScriptOutput = payment({
-      redeem: { output: redeemScript },
-    }).output;
-    if (!scriptPubKey.equals(redeemScriptOutput)) {
-      throw new Error(
-        `${paymentScriptName} for ${ioType} #${inputIndex} doesn't match the scriptPubKey in the prevout`,
-      );
-    }
-  };
-}
-const checkRedeemScript = scriptCheckerFactory(payments.p2sh, 'Redeem script');
-const checkWitnessScript = scriptCheckerFactory(
-  payments.p2wsh,
-  'Witness script',
-);
-function getTxCacheValue(key, name, inputs, c) {
-  if (!inputs.every(isFinalized))
-    throw new Error(`PSBT must be finalized to calculate ${name}`);
-  if (key === '__FEE_RATE' && c.__FEE_RATE) return c.__FEE_RATE;
-  if (key === '__FEE' && c.__FEE) return c.__FEE;
-  let tx;
-  let mustFinalize = true;
-  if (c.__EXTRACTED_TX) {
-    tx = c.__EXTRACTED_TX;
-    mustFinalize = false;
-  } else {
-    tx = c.__TX.clone();
-  }
-  inputFinalizeGetAmts(inputs, tx, c, mustFinalize);
-  if (key === '__FEE_RATE') return c.__FEE_RATE;
-  else if (key === '__FEE') return c.__FEE;
-}
-function getFinalScripts(inputIndex, input, script, isSegwit, isP2SH, isP2WSH) {
-  const scriptType = classifyScript(script);
-  if (!canFinalize(input, script, scriptType))
-    throw new Error(`Can not finalize input #${inputIndex}`);
-  return prepareFinalScripts(
-    script,
-    scriptType,
-    input.partialSig,
-    isSegwit,
-    isP2SH,
-    isP2WSH,
-  );
-}
-function prepareFinalScripts(
-  script,
-  scriptType,
-  partialSig,
-  isSegwit,
-  isP2SH,
-  isP2WSH,
-) {
-  let finalScriptSig;
-  let finalScriptWitness;
-  // Wow, the payments API is very handy
-  const payment = getPayment(script, scriptType, partialSig);
-  const p2wsh = !isP2WSH ? null : payments.p2wsh({ redeem: payment });
-  const p2sh = !isP2SH ? null : payments.p2sh({ redeem: p2wsh || payment });
-  if (isSegwit) {
-    if (p2wsh) {
-      finalScriptWitness = (0, psbtutils_1.witnessStackToScriptWitness)(
-        p2wsh.witness,
-      );
-    } else {
-      finalScriptWitness = (0, psbtutils_1.witnessStackToScriptWitness)(
-        payment.witness,
-      );
-    }
-    if (p2sh) {
-      finalScriptSig = p2sh.input;
-    }
-  } else {
-    if (p2sh) {
-      finalScriptSig = p2sh.input;
-    } else {
-      finalScriptSig = payment.input;
-    }
-  }
-  return {
-    finalScriptSig,
-    finalScriptWitness,
-  };
-}
-function getHashAndSighashType(
-  inputs,
-  inputIndex,
-  pubkey,
-  cache,
-  sighashTypes,
-) {
-  const input = (0, utils_1.checkForInput)(inputs, inputIndex);
-  const { hash, sighashType, script } = getHashForSig(
-    inputIndex,
-    input,
-    cache,
-    false,
-    sighashTypes,
-  );
-  checkScriptForPubkey(pubkey, script, 'sign');
-  return {
-    hash,
-    sighashType,
-  };
-}
-function getHashForSig(inputIndex, input, cache, forValidate, sighashTypes) {
-  const unsignedTx = cache.__TX;
-  const sighashType =
-    input.sighashType || transaction_1.Transaction.SIGHASH_ALL;
-  checkSighashTypeAllowed(sighashType, sighashTypes);
-  let hash;
-  let prevout;
-  if (input.nonWitnessUtxo) {
-    const nonWitnessUtxoTx = nonWitnessUtxoTxFromCache(
-      cache,
-      input,
-      inputIndex,
-    );
-    const prevoutHash = unsignedTx.ins[inputIndex].hash;
-    const utxoHash = nonWitnessUtxoTx.getHash();
-    // If a non-witness UTXO is provided, its hash must match the hash specified in the prevout
-    if (!prevoutHash.equals(utxoHash)) {
-      throw new Error(
-        `Non-witness UTXO hash for input #${inputIndex} doesn't match the hash specified in the prevout`,
-      );
-    }
-    const prevoutIndex = unsignedTx.ins[inputIndex].index;
-    prevout = nonWitnessUtxoTx.outs[prevoutIndex];
-  } else if (input.witnessUtxo) {
-    prevout = input.witnessUtxo;
-  } else {
-    throw new Error('Need a Utxo input item for signing');
-  }
-  const { meaningfulScript, type } = getMeaningfulScript(
-    prevout.script,
-    inputIndex,
-    'input',
-    input.redeemScript,
-    input.witnessScript,
-  );
-  if (['p2sh-p2wsh', 'p2wsh'].indexOf(type) >= 0) {
-    hash = unsignedTx.hashForWitnessV0(
-      inputIndex,
-      meaningfulScript,
-      prevout.value,
-      sighashType,
-    );
-  } else if ((0, psbtutils_1.isP2WPKH)(meaningfulScript)) {
-    // P2WPKH uses the P2PKH template for prevoutScript when signing
-    const signingScript = payments.p2pkh({
-      hash: meaningfulScript.slice(2),
-    }).output;
-    hash = unsignedTx.hashForWitnessV0(
-      inputIndex,
-      signingScript,
-      prevout.value,
-      sighashType,
-    );
-  } else {
-    // non-segwit
-    if (
-      input.nonWitnessUtxo === undefined &&
-      cache.__UNSAFE_SIGN_NONSEGWIT === false
-    )
-      throw new Error(
-        `Input #${inputIndex} has witnessUtxo but non-segwit script: ` +
-          `${meaningfulScript.toString('hex')}`,
-      );
-    if (!forValidate && cache.__UNSAFE_SIGN_NONSEGWIT !== false)
-      console.warn(
-        'Warning: Signing non-segwit inputs without the full parent transaction ' +
-          'means there is a chance that a miner could feed you incorrect information ' +
-          "to trick you into paying large fees. This behavior is the same as Psbt's predecessor " +
-          '(TransactionBuilder - now removed) when signing non-segwit scripts. You are not ' +
-          'able to export this Psbt with toBuffer|toBase64|toHex since it is not ' +
-          'BIP174 compliant.\n*********************\nPROCEED WITH CAUTION!\n' +
-          '*********************',
-      );
-    hash = unsignedTx.hashForSignature(
-      inputIndex,
-      meaningfulScript,
-      sighashType,
-    );
-  }
-  return {
-    script: meaningfulScript,
-    sighashType,
-    hash,
-  };
-}
-function getAllTaprootHashesForSig(inputIndex, input, inputs, cache) {
-  const allPublicKeys = [];
-  if (input.tapInternalKey) {
-    const key = getPrevoutTaprootKey(inputIndex, input, cache);
-    if (key) {
-      allPublicKeys.push(key);
-    }
-  }
-  if (input.tapScriptSig) {
-    const tapScriptPubkeys = input.tapScriptSig.map(tss => tss.pubkey);
-    allPublicKeys.push(...tapScriptPubkeys);
-  }
-  const allHashes = allPublicKeys.map(pubicKey =>
-    getTaprootHashesForSig(inputIndex, input, inputs, pubicKey, cache),
-  );
-  return allHashes.flat();
-}
-function getPrevoutTaprootKey(inputIndex, input, cache) {
-  const { script } = getScriptAndAmountFromUtxo(inputIndex, input, cache);
-  return (0, psbtutils_1.isP2TR)(script) ? script.subarray(2, 34) : null;
-}
 function trimTaprootSig(signature) {
   return signature.length === 64 ? signature : signature.subarray(0, 64);
-}
-function getTaprootHashesForSig(
-  inputIndex,
-  input,
-  inputs,
-  pubkey,
-  cache,
-  tapLeafHashToSign,
-  allowedSighashTypes,
-) {
-  const unsignedTx = cache.__TX;
-  const sighashType =
-    input.sighashType || transaction_1.Transaction.SIGHASH_DEFAULT;
-  checkSighashTypeAllowed(sighashType, allowedSighashTypes);
-  const prevOuts = inputs.map((i, index) =>
-    getScriptAndAmountFromUtxo(index, i, cache),
-  );
-  const signingScripts = prevOuts.map(o => o.script);
-  const values = prevOuts.map(o => o.value);
-  const hashes = [];
-  if (input.tapInternalKey && !tapLeafHashToSign) {
-    const outputKey =
-      getPrevoutTaprootKey(inputIndex, input, cache) || Buffer.from([]);
-    if ((0, bip371_1.toXOnly)(pubkey).equals(outputKey)) {
-      const tapKeyHash = unsignedTx.hashForWitnessV1(
-        inputIndex,
-        signingScripts,
-        values,
-        sighashType,
-      );
-      hashes.push({ pubkey, hash: tapKeyHash });
-    }
-  }
-  const tapLeafHashes = (input.tapLeafScript || [])
-    .filter(tapLeaf => (0, psbtutils_1.pubkeyInScript)(pubkey, tapLeaf.script))
-    .map(tapLeaf => {
-      const hash = (0, bip341_1.tapleafHash)({
-        output: tapLeaf.script,
-        version: tapLeaf.leafVersion,
-      });
-      return Object.assign({ hash }, tapLeaf);
-    })
-    .filter(
-      tapLeaf => !tapLeafHashToSign || tapLeafHashToSign.equals(tapLeaf.hash),
-    )
-    .map(tapLeaf => {
-      const tapScriptHash = unsignedTx.hashForWitnessV1(
-        inputIndex,
-        signingScripts,
-        values,
-        transaction_1.Transaction.SIGHASH_DEFAULT,
-        tapLeaf.hash,
-      );
-      return {
-        pubkey,
-        hash: tapScriptHash,
-        leafHash: tapLeaf.hash,
-      };
-    });
-  return hashes.concat(tapLeafHashes);
-}
-function checkSighashTypeAllowed(sighashType, sighashTypes) {
-  if (sighashTypes && sighashTypes.indexOf(sighashType) < 0) {
-    const str = sighashTypeToString(sighashType);
-    throw new Error(
-      `Sighash type is not allowed. Retry the sign method passing the ` +
-        `sighashTypes array of whitelisted types. Sighash type: ${str}`,
-    );
-  }
-}
-function getPayment(script, scriptType, partialSig) {
-  let payment;
-  switch (scriptType) {
-    case 'multisig':
-      const sigs = getSortedSigs(script, partialSig);
-      payment = payments.p2ms({
-        output: script,
-        signatures: sigs,
-      });
-      break;
-    case 'pubkey':
-      payment = payments.p2pk({
-        output: script,
-        signature: partialSig[0].signature,
-      });
-      break;
-    case 'pubkeyhash':
-      payment = payments.p2pkh({
-        output: script,
-        pubkey: partialSig[0].pubkey,
-        signature: partialSig[0].signature,
-      });
-      break;
-    case 'witnesspubkeyhash':
-      payment = payments.p2wpkh({
-        output: script,
-        pubkey: partialSig[0].pubkey,
-        signature: partialSig[0].signature,
-      });
-      break;
-  }
-  return payment;
 }
 function getScriptFromInput(inputIndex, input, cache) {
   const unsignedTx = cache.__TX;
@@ -1367,7 +995,7 @@ function getScriptFromInput(inputIndex, input, cache) {
     res.script = input.redeemScript;
   } else {
     if (input.nonWitnessUtxo) {
-      const nonWitnessUtxoTx = nonWitnessUtxoTxFromCache(
+      const nonWitnessUtxoTx = (0, cache_1.nonWitnessUtxoTxFromCache)(
         cache,
         input,
         inputIndex,
@@ -1411,165 +1039,14 @@ function getSignersFromHD(inputIndex, inputs, hdKeyPair) {
   });
   return signers;
 }
-function getSortedSigs(script, partialSig) {
-  const p2ms = payments.p2ms({ output: script });
-  // for each pubkey in order of p2ms script
-  return p2ms.pubkeys
-    .map(pk => {
-      // filter partialSig array by pubkey being equal
-      return (
-        partialSig.filter(ps => {
-          return ps.pubkey.equals(pk);
-        })[0] || {}
-      ).signature;
-      // Any pubkey without a match will return undefined
-      // this last filter removes all the undefined items in the array.
-    })
-    .filter(v => !!v);
-}
-function scriptWitnessToWitnessStack(buffer) {
-  let offset = 0;
-  function readSlice(n) {
-    offset += n;
-    return buffer.slice(offset - n, offset);
-  }
-  function readVarInt() {
-    const vi = varuint.decode(buffer, offset);
-    offset += varuint.decode.bytes;
-    return vi;
-  }
-  function readVarSlice() {
-    return readSlice(readVarInt());
-  }
-  function readVector() {
-    const count = readVarInt();
-    const vector = [];
-    for (let i = 0; i < count; i++) vector.push(readVarSlice());
-    return vector;
-  }
-  return readVector();
-}
-function sighashTypeToString(sighashType) {
-  let text =
-    sighashType & transaction_1.Transaction.SIGHASH_ANYONECANPAY
-      ? 'SIGHASH_ANYONECANPAY | '
-      : '';
-  const sigMod = sighashType & 0x1f;
-  switch (sigMod) {
-    case transaction_1.Transaction.SIGHASH_ALL:
-      text += 'SIGHASH_ALL';
-      break;
-    case transaction_1.Transaction.SIGHASH_SINGLE:
-      text += 'SIGHASH_SINGLE';
-      break;
-    case transaction_1.Transaction.SIGHASH_NONE:
-      text += 'SIGHASH_NONE';
-      break;
-  }
-  return text;
-}
-function addNonWitnessTxCache(cache, input, inputIndex) {
-  cache.__NON_WITNESS_UTXO_BUF_CACHE[inputIndex] = input.nonWitnessUtxo;
-  const tx = transaction_1.Transaction.fromBuffer(input.nonWitnessUtxo);
-  cache.__NON_WITNESS_UTXO_TX_CACHE[inputIndex] = tx;
-  const self = cache;
-  const selfIndex = inputIndex;
-  delete input.nonWitnessUtxo;
-  Object.defineProperty(input, 'nonWitnessUtxo', {
-    enumerable: true,
-    get() {
-      const buf = self.__NON_WITNESS_UTXO_BUF_CACHE[selfIndex];
-      const txCache = self.__NON_WITNESS_UTXO_TX_CACHE[selfIndex];
-      if (buf !== undefined) {
-        return buf;
-      } else {
-        const newBuf = txCache.toBuffer();
-        self.__NON_WITNESS_UTXO_BUF_CACHE[selfIndex] = newBuf;
-        return newBuf;
-      }
-    },
-    set(data) {
-      self.__NON_WITNESS_UTXO_BUF_CACHE[selfIndex] = data;
-    },
-  });
-}
-function inputFinalizeGetAmts(inputs, tx, cache, mustFinalize) {
-  let inputAmount = 0;
-  inputs.forEach((input, idx) => {
-    if (mustFinalize && input.finalScriptSig)
-      tx.ins[idx].script = input.finalScriptSig;
-    if (mustFinalize && input.finalScriptWitness) {
-      tx.ins[idx].witness = scriptWitnessToWitnessStack(
-        input.finalScriptWitness,
-      );
-    }
-    if (input.witnessUtxo) {
-      inputAmount += input.witnessUtxo.value;
-    } else if (input.nonWitnessUtxo) {
-      const nwTx = nonWitnessUtxoTxFromCache(cache, input, idx);
-      const vout = tx.ins[idx].index;
-      const out = nwTx.outs[vout];
-      inputAmount += out.value;
-    }
-  });
-  const outputAmount = tx.outs.reduce((total, o) => total + o.value, 0);
-  const fee = inputAmount - outputAmount;
-  if (fee < 0) {
-    throw new Error('Outputs are spending more than Inputs');
-  }
-  const bytes = tx.virtualSize();
-  cache.__FEE = fee;
-  cache.__EXTRACTED_TX = tx;
-  cache.__FEE_RATE = Math.floor(fee / bytes);
-}
-function nonWitnessUtxoTxFromCache(cache, input, inputIndex) {
-  const c = cache.__NON_WITNESS_UTXO_TX_CACHE;
-  if (!c[inputIndex]) {
-    addNonWitnessTxCache(cache, input, inputIndex);
-  }
-  return c[inputIndex];
-}
-function getScriptFromUtxo(inputIndex, input, cache) {
-  const { script } = getScriptAndAmountFromUtxo(inputIndex, input, cache);
-  return script;
-}
-function getScriptAndAmountFromUtxo(inputIndex, input, cache) {
-  if (input.witnessUtxo !== undefined) {
-    return {
-      script: input.witnessUtxo.script,
-      value: input.witnessUtxo.value,
-    };
-  } else if (input.nonWitnessUtxo !== undefined) {
-    const nonWitnessUtxoTx = nonWitnessUtxoTxFromCache(
-      cache,
-      input,
-      inputIndex,
-    );
-    const o = nonWitnessUtxoTx.outs[cache.__TX.ins[inputIndex].index];
-    return { script: o.script, value: o.value };
-  } else {
-    throw new Error("Can't find pubkey in input without Utxo data");
-  }
-}
 function pubkeyInInput(pubkey, input, inputIndex, cache) {
-  const script = getScriptFromUtxo(inputIndex, input, cache);
-  const { meaningfulScript } = getMeaningfulScript(
+  const script = (0, script_1.getScriptFromUtxo)(inputIndex, input, cache);
+  const { meaningfulScript } = (0, script_1.getMeaningfulScript)(
     script,
     inputIndex,
     'input',
     input.redeemScript,
     input.witnessScript,
-  );
-  return (0, psbtutils_1.pubkeyInScript)(pubkey, meaningfulScript);
-}
-function pubkeyInOutput(pubkey, output, outputIndex, cache) {
-  const script = cache.__TX.outs[outputIndex].script;
-  const { meaningfulScript } = getMeaningfulScript(
-    script,
-    outputIndex,
-    'output',
-    output.redeemScript,
-    output.witnessScript,
   );
   return (0, psbtutils_1.pubkeyInScript)(pubkey, meaningfulScript);
 }
@@ -1590,86 +1067,18 @@ function redeemFromFinalScriptSig(finalScript) {
 }
 function redeemFromFinalWitnessScript(finalScript) {
   if (!finalScript) return;
-  const decomp = scriptWitnessToWitnessStack(finalScript);
+  const decomp = (0, script_1.scriptWitnessToWitnessStack)(finalScript);
   const lastItem = decomp[decomp.length - 1];
   if (isPubkeyLike(lastItem)) return;
   const sDecomp = bscript.decompile(lastItem);
   if (!sDecomp) return;
   return lastItem;
 }
-function compressPubkey(pubkey) {
-  if (pubkey.length === 65) {
-    const parity = pubkey[64] & 1;
-    const newKey = pubkey.slice(0, 33);
-    newKey[0] = 2 | parity;
-    return newKey;
-  }
-  return pubkey.slice();
-}
 function isPubkeyLike(buf) {
   return buf.length === 33 && bscript.isCanonicalPubKey(buf);
 }
 function isSigLike(buf) {
   return bscript.isCanonicalScriptSignature(buf);
-}
-function getMeaningfulScript(
-  script,
-  index,
-  ioType,
-  redeemScript,
-  witnessScript,
-) {
-  const isP2SH = (0, psbtutils_1.isP2SHScript)(script);
-  const isP2SHP2WSH =
-    isP2SH && redeemScript && (0, psbtutils_1.isP2WSHScript)(redeemScript);
-  const isP2WSH = (0, psbtutils_1.isP2WSHScript)(script);
-  if (isP2SH && redeemScript === undefined)
-    throw new Error('scriptPubkey is P2SH but redeemScript missing');
-  if ((isP2WSH || isP2SHP2WSH) && witnessScript === undefined)
-    throw new Error(
-      'scriptPubkey or redeemScript is P2WSH but witnessScript missing',
-    );
-  let meaningfulScript;
-  if (isP2SHP2WSH) {
-    meaningfulScript = witnessScript;
-    checkRedeemScript(index, script, redeemScript, ioType);
-    checkWitnessScript(index, redeemScript, witnessScript, ioType);
-    checkInvalidP2WSH(meaningfulScript);
-  } else if (isP2WSH) {
-    meaningfulScript = witnessScript;
-    checkWitnessScript(index, script, witnessScript, ioType);
-    checkInvalidP2WSH(meaningfulScript);
-  } else if (isP2SH) {
-    meaningfulScript = redeemScript;
-    checkRedeemScript(index, script, redeemScript, ioType);
-  } else {
-    meaningfulScript = script;
-  }
-  return {
-    meaningfulScript,
-    type: isP2SHP2WSH
-      ? 'p2sh-p2wsh'
-      : isP2SH
-      ? 'p2sh'
-      : isP2WSH
-      ? 'p2wsh'
-      : 'raw',
-  };
-}
-function checkInvalidP2WSH(script) {
-  if (
-    (0, psbtutils_1.isP2WPKH)(script) ||
-    (0, psbtutils_1.isP2SHScript)(script)
-  ) {
-    throw new Error('P2WPKH or P2SH can not be contained within P2WSH');
-  }
-}
-function classifyScript(script) {
-  if ((0, psbtutils_1.isP2WPKH)(script)) return 'witnesspubkeyhash';
-  if ((0, psbtutils_1.isP2PKH)(script)) return 'pubkeyhash';
-  if ((0, psbtutils_1.isP2MS)(script)) return 'multisig';
-  if ((0, psbtutils_1.isP2PK)(script)) return 'pubkey';
-  return 'nonstandard';
 }
 function range(n) {
   return [...Array(n).keys()];
