+++ src/psbt/global/cache.js
@@ -1,0 +1,79 @@
+'use strict';
+Object.defineProperty(exports, '__esModule', { value: true });
+exports.addNonWitnessTxCache =
+  exports.nonWitnessUtxoTxFromCache =
+  exports.getTxCacheValue =
+  exports.checkTxInputCache =
+  exports.checkCache =
+    void 0;
+const bufferutils_1 = require('../../bufferutils');
+const transaction_1 = require('../../transaction');
+const _1 = require('.');
+const input_1 = require('../input');
+function checkCache(cache) {
+  if (cache.__UNSAFE_SIGN_NONSEGWIT !== false) {
+    throw new Error('Not BIP174 compliant, can not export');
+  }
+}
+exports.checkCache = checkCache;
+function checkTxInputCache(cache, input) {
+  const key =
+    (0, bufferutils_1.reverseBuffer)(Buffer.from(input.hash)).toString('hex') +
+    ':' +
+    input.index;
+  if (cache.__TX_IN_CACHE[key]) throw new Error('Duplicate input detected.');
+  cache.__TX_IN_CACHE[key] = 1;
+}
+exports.checkTxInputCache = checkTxInputCache;
+function getTxCacheValue(key, name, inputs, c) {
+  if (!inputs.every(_1.isFinalized))
+    throw new Error(`PSBT must be finalized to calculate ${name}`);
+  if (key === '__FEE_RATE' && c.__FEE_RATE) return c.__FEE_RATE;
+  if (key === '__FEE' && c.__FEE) return c.__FEE;
+  let tx;
+  let mustFinalize = true;
+  if (c.__EXTRACTED_TX) {
+    tx = c.__EXTRACTED_TX;
+    mustFinalize = false;
+  } else {
+    tx = c.__TX.clone();
+  }
+  (0, input_1.inputFinalizeGetAmts)(inputs, tx, c, mustFinalize);
+  if (key === '__FEE_RATE') return c.__FEE_RATE;
+  else if (key === '__FEE') return c.__FEE;
+}
+exports.getTxCacheValue = getTxCacheValue;
+function nonWitnessUtxoTxFromCache(cache, input, inputIndex) {
+  const c = cache.__NON_WITNESS_UTXO_TX_CACHE;
+  if (!c[inputIndex]) {
+    addNonWitnessTxCache(cache, input, inputIndex);
+  }
+  return c[inputIndex];
+}
+exports.nonWitnessUtxoTxFromCache = nonWitnessUtxoTxFromCache;
+function addNonWitnessTxCache(cache, input, inputIndex) {
+  cache.__NON_WITNESS_UTXO_BUF_CACHE[inputIndex] = input.nonWitnessUtxo;
+  const tx = transaction_1.Transaction.fromBuffer(input.nonWitnessUtxo);
+  cache.__NON_WITNESS_UTXO_TX_CACHE[inputIndex] = tx;
+  const self = cache;
+  const selfIndex = inputIndex;
+  delete input.nonWitnessUtxo;
+  Object.defineProperty(input, 'nonWitnessUtxo', {
+    enumerable: true,
+    get() {
+      const buf = self.__NON_WITNESS_UTXO_BUF_CACHE[selfIndex];
+      const txCache = self.__NON_WITNESS_UTXO_TX_CACHE[selfIndex];
+      if (buf !== undefined) {
+        return buf;
+      } else {
+        const newBuf = txCache.toBuffer();
+        self.__NON_WITNESS_UTXO_BUF_CACHE[selfIndex] = newBuf;
+        return newBuf;
+      }
+    },
+    set(data) {
+      self.__NON_WITNESS_UTXO_BUF_CACHE[selfIndex] = data;
+    },
+  });
+}
+exports.addNonWitnessTxCache = addNonWitnessTxCache;
